Use stable git instead of unstable overlay
Since we're using an overlay for nixpkgs-unstable, everything in the
overlay will become the canonical version for the whole package set.
Many packages depend on git, so this causes all of those packages to
MISS the cache and sends darwin-rebuilds careening out of control.

The version of git we wanted is now in stable, so we should just use
that.